requirements.txt
# 코어 패키지
Django>=5.1.0,<5.2.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
Markdown>=3.4.0
python-dotenv>=1.0.0

# 배포 패키지
gunicorn>=20.1.0
dj-database-url>=1.0.0

# 테스트 및 CI/CD 패키지
coverage>=7.0.0 